# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal search and RAG tool — indexes data from multiple sources and provides unified search.",
        "title": "Nexus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/connectors/{id}/auth/code": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Completes Telegram authentication by submitting the code received in the Telegram app.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connectors"
                ],
                "summary": "Submit Telegram auth code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connector UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Auth code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
        "/connectors/{id}/auth/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Initiates the Telegram MTProto auth flow. Sends a code to the user's Telegram app.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connectors"
                ],
                "summary": "Start Telegram authentication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connector UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by either a JWT from /auth/login or a personal access token (nexus_pat_...) from /tokens. The \"Bearer \" prefix is required; it is sent as the Authorization header verbatim.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}